+++ src/main/java/io/games/play_book/game_buy_in/domain/GameBuyIn.java
@@ -30,7 +30,7 @@ public class GameBuyIn {
       allocationSize = 1,
       initialValue = 10000)
   @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "primary_sequence")
-  private Integer gameBuyInId;
+  private long gameBuyInId;
 
   @Column(nullable = false, precision = 16, scale = 2)
   private BigDecimal buyInAmount;
