+++ src/main/java/io/games/play_book/game_buy_in/model/GameBuyInDTO.java
@@ -12,14 +12,14 @@ import lombok.Setter;
 @Setter
 public class GameBuyInDTO {
 
-  private Integer gameBuyInId;
+  private Long gameBuyInId;
 
   @NotNull
   @Digits(integer = 16, fraction = 2)
   @JsonFormat(shape = JsonFormat.Shape.STRING)
   private BigDecimal buyInAmount;
 
-  @NotNull private Integer game;
+  @NotNull private Long game;
 
-  @NotNull private Integer seasonPlayer;
+  @NotNull private Long seasonPlayer;
 }
